FEAT: Add iterators to elementdata (and test)
Allow "for of" loops on ElementData using the AspectDataDictionary or ElementDataDictionary classes.

We can change how this iterator works but for now I've made it just output the keys as that makes more sense than values as you wouldn't know what you're getting.

Need to patch this into SwigData on DeviceDetection too where it'll be a bit more complex but easy enough (look at size and iterate over indexes).

```javascript
for (property in flowData.device){
  console.log(flowData.device[property].value
}
```