FEAT: New Cloud Engine base class, ready() promise on FlowElements
- Made a Cloud Engine base class and much improved the properties and evidencekey filters of cloud engines extended from it. Any new Cloud Engine can now just extend from this class and add a new dataKey.
- Made a new ready() Promise on all FlowElements that if they override, can be used to check if a FlowElement is ready such as if the properties and evidenceKeyFilters on cloud engines have been populated

I've kept the old way properties were passed through via the elementdata of the cloud request engine to make this easier to merge in and keep compatibility with the old cloud engines. But we should patch all the other cloud engines to just extend from the base class as it means it's a couple of lines of code to make a new cloud engine and far easier to maintain.

I've put in a little test.js file which shows the new way of making a device engine and tests it's all ok but that should be removed before merging.