FEAT: Error handling on bad resource keys etc
This one took a bit of figuring out but got there. tl;dr = example at bottom of this.

The problem is we need an async error on the constructor of the cloud request engine. But async constructors are (though possible because you can return anything from a constructor in JavaScript as discovered elsewhere in this project) frowned upon (checked a few places, they all said don't do it) so I've had to use two ways of recording these errors to catch everything:

If you haven't attached the CloudRequestEngine to a Pipeline, the error will be in the .ready() promise that checks if the engine is ready. This only really makes sense on the CloudRequestEngine itself rather than the CloudEngine that depends on it as the latter without the former doesn't work at all but I've put it on both.

The more sensible way is that the async function called by the CloudRequestEngine constructor now loops over any pipelines it is in and logs out any errors from the cloud API to them.

So pipeline.on("error") will catch it. And if you don't have pipeline.on("error") you'll get a crash.

Later on in the pipeline (if you've handled the error as above), you will after processing get the error (in FlowData.errors) that has come from the cloud API once keyed under CloudRequestEngine and then keyed again under each of the cloud engines.

I put in an extra commit at the start which was the property database in pipeline.js crashing out when there were no properties but some were expected (like the failed constructor). Handling the constructor errors means this shouldn't be an issue anymore but it might be for something else so keeping it in.

Here's a quick demo you can run.

```JavaScript

let Cloud = require("./index");
let Core = require("fiftyone.pipeline.core");

let deviceEngine = new Cloud.CloudEngine();
deviceEngine.dataKey = "device";

let lonelyCRE = new Cloud.CloudRequestEngine({resourceKey:"not-a-resource-key"});

lonelyCRE.ready().catch(function(error){

  console.log("Error from .ready()", error);

});

let pipeline = new Core.PipelineBuilder()
  .add(new Cloud.CloudRequestEngine({
    resourceKey: "not-a-resource-key"
  }))
  .add(deviceEngine)
  .build();

// This'll catch all the errors, including the constructor errors
pipeline.on("error", function(error){

  console.log("Error from pipeline logging", error);

});

let flowData = pipeline.createFlowData();

flowData.process().then(function(){

  // This'll show errors for device and cloud (bad resource key)
  console.log("FlowData error object", flowData.errors);

})

```