Merged PR 3387: FEAT/OPTIM/TEST/EXAMPLE/BUG/BUILD: Updated the error messages to provide more helpful information
Changes:

OPTIM: Updated the error messages to provide more helpful information.
TEST: Check that correct error message are returned for each failed scenario.
OPTIM: Accessing invalid 'element' using dot notation will now throw exception with appropriate message.
TEST: Check that exception is thrown and correct error message is returned when accessing invalid element using dot notation.
BUG: Correctly checked for 'addJavaScriptBuilder' option in 'pipelineBuilder.js'.
FEAT: ElementData now return more helpful error message when accessing invalid properties.
TEST: Check that helpful error message is returned when accessing invalid properties.
EXAMPLE: Updated onPremiseFlowElement example to handle 'favicon.ico' request.
BUILD: Removed build and test for Node 6 and 8.

Failure scenarios - some are error messages and some are tests added or both:
If an expected property is not present in the data and the flow element is a cloud engine, then let the user know that it's probably due to the resource key.
Access a property that was not in the list of desired properties passed to the engine during construction
Attempt to access the value for a property that is not present in the dataset, and where there is no engine-specific handling
Attempt to access an element data that does not exist against the flow data. For example, flowdata.location when there is only a device detection engine.
Attempt to access a property value under a cloud engine. The resource key does not have any properties under that product (e.g. the key has no location properties and the caller is trying to get the value of location.country)
Attempt to access a property value under a cloud engine. The resource key does not include the requested property, but does include other properties under the same engine.
Attempt to access a property value under a cloud engine and neither of the above scenarios is correct.
Resource key is not valid.